+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgrade the `org.realityforge.zemeckis` artifact to version `0.09`.
 * Upgrade the `org.realityforge.grim` artifacts to version `0.05`.
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
+* Migrate from using Elemental2 to Akasha when interacting with the Browser API.
 
 ### [v6.105](https://github.com/replicant4j/replicant/tree/v6.105) (2021-02-01) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.104...v6.105)